requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-bootstrap-static==4.2.1
django-debug-toolbar==2.2
django-silk==4.0.1
gprof2dot==2019.11.30
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pycodestyle==2.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
